converted ActionsCard to container and seperated table from card
